# terraform/gcp/modules/fhir-converter/api-spec.yaml
swagger: '2.0'
info:
  title: FHIR Converter API
  description: API for converting healthcare data formats to FHIR
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
paths:
...
  /convert/fhir-to-hl7v2:
    post:
      summary: Convert FHIR to HL7v2
      operationId: convertFhirToHl7v2
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              fhir_resource:
                type: object
                description: FHIR resource
              template_collection:
                type: string
                description: Template collection name
      responses:
        '200':
          description: Conversion successful
          schema:
            type: object
            properties:
              hl7v2_message:
                type: string
                description: HL7v2 message
        '400':
          description: Invalid input
        '500':
          description: Conversion error
...